+++ packages/ui-helpers/src/index.tsx
@@ -30,3 +30,4 @@ export { BarContainer } from "./votingBar/BarContainer";
 export { Header } from "./Header";
 export { InformationBanner } from "./InformationBanner";
 export { ConsentModal } from "./termsOfServices/ConsentModal";
+export { SystemErrorBanner } from "./SystemErrorBanner";